리팩토링 - 게시판 목록 불러오기완료
